debug doesn't need to be postprocessed
