+++ sdk-core-protos/Cargo.toml
@@ -30,6 +30,7 @@ uuid = { version = "1.1", features = ["v4"], optional = true }
 
 [build-dependencies]
 tonic-build = { workspace = true }
+prost-build = "0.13"
 prost-wkt-build = "0.6"
 
 [lints]
